Test build digest calculation with circular dependencies
It's not enough to just have circular dependency for `spk build` to
calculate a different build digest than `spk test`. There has to be a
mutual 3rd dependency shared between the two packages involved in a
circular dependency. Then, `spk test` still arguably calculates the
"correct" build digest, but `spk build` has extra option values that come
from the solution when it shouldn't be using that to derive a build digest.

Signed-off-by: J Robert Ray <jrray@jrray.org>