Add explicit permissions for created fs objects
By not setting the default permissions, we are relying on the active
umask which may create databases which are not able to be shared between
users as we typically aim to support. We also have a pseudo contract
already defined by the FSHashStore::file_permissions field which
signifies that these permissions should be used for created files.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>