Add "last" arg marker to spfs-join
Clap has a limitation in how it can support alternative args. For example,
in spfs-join, we intend to support two different ways of specifying the
runtime, and Clap's generated usage reflects this:

    Usage: spfs-join [OPTIONS] <RUNTIME|--pid <PID>> [<COMMAND>] [ARGS]...
                               ^^^^^^^^^^^^^^^^^^^^^
Unfortunately, even when `--pid <pid>` is present on the command line, it
still tries to assign any additional positional arguments into the
"runtime" position, as in:

    $ spfs join --pid 162509 -- echo hi
    error: The argument '--pid <PID>' cannot be used with '<RUNTIME>'

    USAGE:
        spfs-join <RUNTIME|--pid <PID>> <COMMAND>

This happens with or without adding `--` to the command line.

By adding "last", Clap will _require_ the command line contains `--` if any
command is being provided, but it removes the parsing ambiguity and makes
the command line shown above work correctly.

Signed-off-by: J Robert Ray <jrray@jrray.org>