Reverse the application of overlayfs index option
As it turns out, I misunderstood how the index=on option works. Some
more clear documentation (thanks arch) points out that breaking hard
links is acutally the default behavior. We could consider removing this
option alltogether, but I wanted to submit a simple change for now to
resolve current issues with the option.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>