Fix spk export failing if builds are missing
In our origin repo there are some traces of a package that may have been
partually published, where there are "spec build" tags but no "pkg build"
tags, as in:

    spfs/tags/spk/spec/pkg-name/1.0.0/6XFZP6SR.tag (exists)
    spfs/tags/spk/pkg/pkg-name (does not exist)

A _new_ build of this package failed to `spk export` due to these remnants
found in the origin repo:

     INFO exporting pkg=pkg-name/1.0.0
     INFO exporting pkg=pkg-name/1.0.0/src
     INFO exporting pkg=pkg-name/1.0.0/5MDZ2ADM
     WARN Ensure that you are specifying at least a package and
     WARN version number when exporting from the local repository
    ERROR Package not found: pkg-name/1.0.0/6XFZP6SR

This commit takes a minimally invasive approach of ignoring
PackageNotFoundError when these are generated about specific builds, as in
builds that exist as far as `list_package_builds` is concerned, but fail to
be loaded. It will still hard error if one attempts to export a specific
build that does not exist.

We don't think that we should have to manually remove the "spec build" tags
from our origin repo in order for `spk export` to work.

Signed-off-by: J Robert Ray <jrray@imageworks.com>