Fix deadlock when hardlink limit is reached
The render_blob function was acquiring a semaphore, but was also
recursive as it would call itself with a different render type if it
found files that were at the limit for hard links. This recursion would
cause it to need another semaphore and so could eventual cause a
deadlock if a lot of concurrent files all recursed at the same time and
could not progress.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>