Don't lose secondary_repositories in spfs run
When using `spk env <args>`, the first that that happens is an exec of
`spfs run - -- spk env --no-runtime <args>`, meaning that in cmd_run.rs it
will go into the first branch of the if here and it wasn't populating
`runtime.config.secondary_repositories`. Then, later in spfs-enter it would
read the config out of the runtime and have an empty
`secondary_repositories`, meaning that when using fuse w/ spk env, any
objects not copied locally would fail to resolve.

Signed-off-by: J Robert Ray <jrray@imageworks.com>