Support variants introducing packages with components
It has been possible to introduce a new "pkg" dependency in a variant, as
in:

    pkg: demo/1.2.3
    build:
      variants:
        - { "new-pkg": "1.2.3" }

A need came up for the ability to introduce a new "pkg" dependency in a
variant but also request specific components for that dependency. Support
this by being more flexible when parsing the variants map.

    pkg: demo/1.2.3
    build:
      variants:
        - { "new-pkg:{comp1,comp2}": "1.2.3" }

The key of this map must not require the value be a valid option name.
Attempt to parse it as a package name expression first. But some values are
valid option names and not package names, like "python.abi", so it has to
still have a fallback to parsing as an OptName.

Signed-off-by: J Robert Ray <jrray@jrray.org>