Add support for globs when reordering package requests
Create a new type BuildKeyPromotionPatterns to formalize this behavior and
make it easier to test. This contains the list of names that will be
promoted to the front of the request order. Instead of simple string
matching, add support for globs using the glob crate.

Make the order that the patterns are specified in significant and order
the matches based on this.

Add "*platforms*" to the default list of patterns, as needed at SPI.
This still needs to be moved to the spk config file.

Signed-off-by: J Robert Ray <jrray@imageworks.com>