Fix `spk install` failing while using OverlayFsWithFuse
More generally, update find_layer_by_filename/find_layers_by_filename so
they don't literally operate on the "local repo" and can resolve objects
from whatever secondary repositories are active.

It is assumed that this was using the "local repo" not because it
shouldn't access any other repo but because before fuse it was always
possible to find these objects locally.

Signed-off-by: J Robert Ray <jrray@imageworks.com>