Use sccache for coverage.yml
The coverage job is now also failing with out of disk space errors.
Switch to sccache like was done for the rust job, to reduce disk usage
(hopefully!).

Include a spelling correction so there is a change to a source file so
the coverage job will run, to verify this change.

Signed-off-by: J Robert Ray <jrray@jrray.org>