Update to serde_yaml 0.9.x
The 0.9 release is about a year old now and there are some important
looking bug fixes compared to the version we were on.

format_serde_error has gone stale and hasn't been updated to support 0.9,
so workaround that by manually converting serde_yaml::Error.

The error output has changed mostly for the better, but in some cases it
doesn't report the location anymore. It looks like this is happening when
the location used to be reported as "line 1 column 1" and in other cases
it still reports a location.

Changes: Update expected output from du tests (specs are formatted
         differently now).

Signed-off-by: J Robert Ray <jrray@jrray.org>