Support conditional install requirements
While waiting for the next spec format, implement a form of conditional
install requirements by augmenting the `fromBuildEnv` mechanism so it
conditionally doesn't fail if the package is missing in the build
environment.

As demonstrated in this example spec:

```yaml
pkg: top/1.0.0
build:
    variants:
    - { dep1: "1.0.0" }
    - { dep2: "1.0.0" }
install:
    requirements:
    - { pkg: "dep1", fromBuildEnv: true, ifPresentInBuildEnv: true }
    - { pkg: "dep2", fromBuildEnv: true, ifPresentInBuildEnv: true }
```

The two variants introduce two different package requirements. The
install requirements have a new modifier "ifPresentInBuildEnv" which
makes it so if the package named is not found in the build environment,
the requirement is omitted instead of this being an error.

This isn't as flexible as the "v1" spec will allow, but it covers a
commonly requested case of supporting optional dependencies depending on
the variant being built. One common example is packages having different
dependencies depending on if python 2 or 3 is being used.

Signed-off-by: J Robert Ray <jrray@jrray.org>