Adds spk info /spfs/file/path suport and changes spk info to not run a solve
Adds 'spk info /spfs/file/path' support (#690) to show what object(s)
contain a filepath from the current runtime. The spk cmd_bake
functions are moved to allow their reuse (#688), along with the
bake-related component errors that are moved to spk-schema crate.

Updates 'spk info <package>' to run without solving (#691) for the
package. The info command looks up the package, version, build spec
directly from the given package request, without running a full
solve. For backwards compatibility the '--full-solve' option is added
to allow the old solving method to be used. But the expectation is
that this will be removed at some point in the future.

Adds -f format option to spk info for package spec output in yaml or
json (#692)

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>