Add configurable tokio thread limits to spfs
The fuse filesystem and monitor are long-running and were spawning an unreasonable
number of threads on machines with a high core count even if the
filesystem was not being used. This change simply allows a global limit
to be placed in spfs config file as well as establishing a more
reasonable default.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>