Support double underscores for env overrides
The `config` crate's stance on handling env overrides for nested structs
that have field names with underscores is to set the separator to "__",
as in:

    # unambiguously specify spfs::Config::SubStruct::example_field
    export SPFS_SUBSTRUCT__EXAMPLE_FIELD="..."

Support this while maintaining backwards compatibility by adding two
`Environment::with_prefix` sources. The one that looks for "_" is last
so it takes precedence.

Add a test to verify the precedence behavior and confirm backwards
compatibility with the single underscore names. `$SPFS_STORAGE_ROOT`
will still be recognized.

Spk's config has been updated to behave the same way, but **not in a
backward compatible way**. Since the spk config file is relatively new
there is an assumption that there aren't any entrenched use cases that
were depending on the custom env injection found in spk's config.rs.

Using double underscores will handle nested structs of any depth and so
it is a slight win over the custom code that only handles one level of
nesting, and behavior consistency between spfs and spk is important.

Signed-off-by: J Robert Ray <jrray@jrray.org>