Allow a package to modify its own files during build
This resolves #862.

The validation code does not have enough context to allow this special
case; it does not know the what package (if any) owns the file if it
detects a modification. Therefore, use a callback mechanism to allow
validation errors to be filtered/ignored.

On the calling side, inspect the validation errors and skip any that match
our criteria that it is about altering a file and the file belongs to a
package with the same name as the one being built.

Add a test case to verify that it is now possible for a build to modify its
own files, and that if it modifies any other package's files this will
still be detected and fail the build as before.

As a bonus, check the question posed in #862 and test that building a
package in a circular dependency situation is possible even when building
a version that would be incompatible with the other version of itself that
is being brought into the build environment.

Another issue with circular dependencies is that the build may create
new files that overwrite existing files, but these files may have the
same contents as before and not be detected as changed.

These new files are found when walking the upperdir, but they were not
included in the filter and so were being skipped. To solve this, add all
the files that belong to the circularly-included package to the filter.
This doesn't mean that all these files get added to the new package, but
it prevents skipping over the files that were touched during the build.

Rework the filter using a custom type to tailor the behavior to the
need. Stop passing a `&[Diff]` both because when adding the extra files,
we don't want to have to make a `Diff` value for them, when all we need
is a path; and to optimize the poor algorithmic complexity of repeatedly
walking the slice to check paths. Now a HashSet is used for fast
lookups.

A side effect of allowing packages with circular dependencies to modify
their own files is that can include deleting files. If these removals
are allowed to be captured as a mask entry then this poisons the package
for downstream users, who run into MustNotAlterExistingFiles for this
phantom removed file.

Signed-off-by: J Robert Ray <jrray@jrray.org>