Allow a package to modify its own files during build
This resolves #862.

The validation code does not have enough context to allow this special
case; it does not know the what package (if any) owns the file if it
detects a modification. Therefore, it can't stop on the first validation
error it finds, but must return all the validation errors, so at a higher
level these can be inspected and optionally ignored.

Create a new error type so this more complex error can be returned. It
contains all the validation errors found plus it still contains the
calculated diff, in case the caller wishes to continue despite the
validation errors.

On the calling side, inspect the validation errors and skip any that match
our criteria that it is about altering a file and the file belongs to a
package with the same name as the one being built.

Add a test case to verify that it is now possible for a build to modify its
own files, and that if it modifies any other package's files this will
still be detected and fail the build as before.

Signed-off-by: J Robert Ray <jrray@jrray.org>