Allow validation to return multiple errors
Fix the problem with `test_package_with_circular_dep_can_modify_files`
that was making it pass without these other changes. It was resolving
the 1.0.1 build of circ created earlier in the test and so the
"version.txt" file was not actually being altered.

In order to ignore some validation errors while still detecting and
stopping for other validation errors, all the validation errors need to
be collected instead of erroring on the first one.

Signed-off-by: J Robert Ray <jrray@jrray.org>