Ensure a circular dependency collects all files
Another issue with circular dependencies is that the build may create
new files that overwrite existing files, but these files may have the
same contents as before and not be detected as changed.

These new files are found when walking the upperdir, but they were not
included in the filter and so were being skipped. To solve this, add all
the files that belong to the circularly-included package to the filter.
This doesn't mean that all these files get added to the new package, but
it prevents skipping over the files that were touched during the build.

Rework the filter using a custom type to tailor the behavior to the
need. Stop passing a `&[Diff]` both because when adding the extra files,
we don't want to have to make a `Diff` value for them, when all we need
is a path; and to optimize the poor algorithmic complexity of repeatedly
walking the slice to check paths. Now a HashSet is used for fast
lookups.

Signed-off-by: J Robert Ray <jrray@jrray.org>