Replace files_to_layers with custom type
To improve performance during build for finding if a circular dependency
exists and accessing the files belonging to that dependency. This index
maintains two ways of accessing the data and ensures they are updated
together.

The operations possible on this new type are limited to what was needed
by the existing code, and not intended to be a full drop-in replacement
for a HashMap or the like.

Signed-off-by: J Robert Ray <jrray@jrray.org>