Add new Platform type
This offers a convenience for writing the package spec for
"meta-packages" that are used to specify versions of a set of packages.
A typical platform package has an empty build options list, and one or
more install requirements, all in IfAlreadyPresent mode, that describe
the versions of dependencies that target compatibility with some
application execution environment, like a DCC.

The new platform spec type reduces the amount of boilerplate needed
compared to using the `v0/package` recipe format, filling in appropriate
defaults automatically for platforms, as well as providing a way to inherit
from another package. The expectation is that a platform would only
inherit from other platforms, but that is not strictly required.

When a platform is built, it produces an ordinary `v0/package` just like
a "normal" `v0/package` spec would do, and it is treated like any other
package for use by downstream consumers.

Note that in this current design, the BuiltPlatform impls are never actually
used, since platforms get built into `v0/package` packages. But
`BuiltPlatform` needs to exist so enum_dispatch doesn't generate
conflicting impls.

Signed-off-by: J Robert Ray <jrray@jrray.org>