Add lazy filesystem repositories
Just like lazy gRPC repositories, this allows for creating an instance
of a filesystem repository that may or may not exist yet. These
repositories have an inner state that must be 'opened' in order to be
used and can fail later on when they are accessed if this is the case.

The OpenFsRepository type is a new addition which skips the lazy
validation and acts like the previous repository implementation. This
opened type must be used for local filesystem renders and other
runtime-related setup functions that would have be done on the normal
FSRepository before.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>