Replace Entry::default with explicit methods
There was a bug where mkdirs was creating entries that looked liked
files. When building from a filesystem the error gets overwritten with
the appropriate mode later on but it shows up in api-based manifest
construction.

Instead of a Default implementation that obscures the details of what is
being created, use explicit methods that can build specific entry types.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>