Fix path canonicalization on windows
The windows std::fs (and tokio::fs) canonicalize functions call into
the windows kernel which turn everthing in to UNC paths with a leading
"\\?\" (for normal file system paths). These are not handled well in most
of our existing code, and so the dunce crate is used to replace all
canonicalize calls. On windows this crate will remove this leading prefix
when it is safe to do so.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>