Introduce OsError trait to abstract errno codes
These codes are reported differently on windows, and so this trait
will enable two things. First it gives us a place to define generic
functions which can handle platform-specific error code checking for
logical cases that we are interested in (eg is this file not found).
Second, it provides a controlled entrypoint for accessing raw error
codes that we can work to make unsafe - ensuring that we don't check
raw platform-specific error codes anywhere that's not appropriate.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>