Add support for durable upperdirs and durable runtimes (--keep_runtime) (#754)
Adds --keep-runtime argument to spfs run, and spk commands. The
keep_runtime value is stored in the runtime's data and used by spfs
enter and monitor.

Using --keep-runtime will make the runtime use a durable upperdir root
path based on the runtime's name, and will stop it being deleted when
the process exits. The upper root paths are stored in the local repo
in durables directory to keep them together and because they must not
be on NFS for overlayfs.

Durable runtimes can only removed by spfs runtime rm and prune when
'--remove-durable' is given, and deletion includes the removal of
the durable upper root path for a durable runtime.  Removing them calls
'spfs clean' because it needs the capability to ignore permissions for
some of the overlayfs related working directories.

Adds '--rerun' argument to 'spfs run' and 'spfs shell'. '--rerun' and 
'REFERENCE' arguments work together with `--` and command 
and args. Removed the 'ARGS' positional from spfs run.

Updates 'spfs monitor' to reset part of a durable runtime's state so it 
can be reused. Does not work with the "index=on" option when 
mounting overlayfs. That option prevents durable upper paths 
(upper dir and work dir) from being reused when a durable runtime
is rerun.

Adds another status display value to spfs runtime list to show durable
runtimes that aren't being used as "saved", instead of "unknown"

Also adds '--force' argument to 'spfs run' and 'spfs shell'. This allows 
a saved runtime to be rerun, even if it is in a strange state due to 
'spfs monitor'  dying during the runtime's last use.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>