Support combining OverlayFsWithFuse and ProxyRepository
The fuse support makes use of ProxyRepository in order to check for
objects in the local repo first, before checking the remote repo(s).
However, it did not support the situation where the remote repo has been
overridden to be a ProxyRepository itself.

There are two different approaches that were considered. The second way as
described below was selected as the preferred approach.

The first way makes it so when opening a file via fuse, when iterating
over the repos, the repos are "flattened" so that if any of the repos
are a ProxyRepository, its internal repos are iterated over
individually.

The second way is changing ProxyRepository::from_config to flatten any
nested proxy repositories, which would perhaps perform better because this
happens only once instead of on every file open.

Modifying ProxyRepository::from_config likely performs better but only
solves the problem in a narrow way, when from_config is used. It is
still possible to programatically construct a ProxyRepository that nests
another ProxyRepository. The logic in from_config added here does not
handle nested ProxyRepositories either, but that could be added.

The iterator change is a more general solution but could be slower due
to need to make a temporary Vec, though this might get optimized away by
the compiler. Unlike the other approach that ignores FallbackProxy, this
approach really ought to support the same behavior for FallbackProxy
too, but in practice it isn't needed and it is difficult to implement.

Signed-off-by: J Robert Ray <jrray@imageworks.com>