Adds --keep-runtime to spfs edit for changing the runtime into a durable one (#859)
The work of changing the runtime to a durable one is done inside 'spfs
enter --make-durable' because copying the current edits to a new durable 
upper dir requires root permissions (for the overlayfs file mask special 
character devices).

This also changes the spfs-monitor process to reload the runtime data
from the filesystem before cleaning up the runtime. This ensures that
changes from other processes, such as 'spfs edit' or 'spfs enter', are
loaded before 'spfs monitor' decides what to do with an exiting runtime.

This adds the 'rsync' requirement to the spk and spfs specs. And it adds
'OverlayFsUnsupportedOnWindows' error for windows implementations.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>