Workaround hang when running spfs-fuse
If stderr is not detached, the following code will hang indefinitely,
blocked on a read() on a pipe.

    #!/usr/bin/env python

    import subprocess

    env_cmd = [
             "spfs commit platform --tag spfs-hang-test",
          ]

    subprocess.run(env_cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)

To induce the hang, this must be run in a spfs environment using
OverlayFsWithFuse and with a modified runtime, as in:

    env SPFS_FILESYSTEM_BACKEND=OverlayFsWithFuse spfs run -- - bash
    $ touch /spfs/foo

An issue will be created with further details on what has been attempted to
try to solve this problem.

Signed-off-by: J Robert Ray <jrray@imageworks.com>