Stop using `now` when fuse returns file attributes
This causes unusual behavior for build tools. For example, when building
python, a file called `Makefile.pre.in` exists in the source files,
which is turned into `Makefile.pre` during the build. However, since
`Makefile.pre.in` exists in the source package and is served by fuse,
after `Makefile.pre` is created and its mtime is compared to
`Makefile.pre.in`, `Makefile.pre.in` appears to be newer, causing the
build to get stuck in a loop trying to recreate `Makefile.pre` over and
over.

In lieu of performing a real stat on the source object in the spfs repo,
return a fixed time in the past, using roughly the time that the spfs
filesystem was mounted.

This fixed time is also used for the atime, so there is a behavior
change here in that files will no longer appear to have been accessed
recently and will act like `noatime` is enabled.

Signed-off-by: J Robert Ray <jrray@jrray.org>