Fix test_validate_build_changeset_collected
As currently written, `must_collect_all_files` will always return an
error if the given list of files is empty. Because this test provides an
empty list of files, the test passes but for the wrong reason.

In #866, `must_collect_all_files` is changed so there must be at least
one `DiffMode::Added` entry in the provided slice of diffs to generate an
error (note that `must_collect_all_files` filters out any diffs that aren't
`DiffMode::Added` and only validates that all _added_ files are collected).

Since `must_collect_all_files` only cares about `DiffMode::Added` diffs,
start using that type of diff for the test data. This change makes the
test continue to pass after the changes in #866.

Signed-off-by: J Robert Ray <jrray@jrray.org>