Rework the Compatibility type to better categorize them
In order to detect common compatibility problems between different paths
through the solver, it is necessary for the Compatibility type to represent
the different possible compatibility problems as discrete types.

One benefit is it becomes possible to shrink many "TRY" lines down into
a single line when all the builds/versions attempted had the same
compatibility problem.

At this stage the code change is a proof of concept and much of the
detail contained in the "TRY" lines has been lost. Subsequent changes in
this branch will restore the original detail.

Signed-off-by: J Robert Ray <jrray@jrray.org>