Improve error for `spk explain <non-existing file>@stage`
Looking for the cause of:

    # already in an spfs runtime; package file exists
    $ ls
    package.spk.yaml

    # explain gives confusing error
    $ spk explain package.spk.yaml@build
    ERROR (link)

      × Error: expected eof at .spk.yaml

Now the error output looks like this:

    $ spk explain package.spk.yaml@build
    ERROR Unable to find "package.spk.yaml" as a file, or existing package/version recipe in any repo
    ERROR   × parsing package.spk.yaml@build as a filename with stage specifier
      ├─▶ finding package recipe for package.spk.yaml
      ╰─▶  > Please check that file path, or package/version request, is correct

The failure is because `spk explain` creates its own empty runtime, so
from the perspective of the new process, /spfs is empty and
package.spk.yaml doesn't exist. Using `spk explain --no-runtime` is
necessary for it to be able to read this file.

Signed-off-by: J Robert Ray <jrray@imageworks.com>