Remove runtime from spk explain
`spk explain` just runs a solver and has no need for an active runtime.
Stop entering an active runtime and hide all the runtime related flags
from the help output.

Continue to parse all the runtime flags that were valid at the time of
this commit. Because of a quirk demonstrated in #912, developers may be
adding `--no-runtime` in some workflows, and will need to have an
opportunity to safely transition to a newer version of spk that doesn't
require this flag to give expected results.

Signed-off-by: J Robert Ray <jrray@jrray.org>