Merge component requests that come from variants
If a package has a build option asking for some components of a package,
and a variant that asks for additional components of that same package,
those requests should be merged together when building the package.

The built package's build.options are expected to contain all the
options from the variant being built.

Add a test to verify this behavior.

Signed-off-by: J Robert Ray <jrray@jrray.org>