Add hidden --allow-empty flag to spfs-commit
This is needed to construct a test around empty platforms, but the flag
is hidden since it is not expected to be used in normal usage.

Note that the act of creating an empty runtime via `spfs run - -- bash`
causes an empty platform to be committed to storage. So it is possible,
but it wasn't possible via `spfs commit`.

Signed-off-by: J Robert Ray <jrray@jrray.org>