Add test to demonstrate digest collision
An empty platform, something that will get created in storage when
starting an empty runtime, happens to have the same digest as a blob
containing 8 null bytes.

A file with 8 null bytes is something cargo will write in the target
directory under some circumstances.

Therefore, creating an spfs layer out of the target directory of a Rust
project can corrupt the spfs storage, as the blob object does not get
written since the file already exists, and that file is a platform
object.

This test serves to demonstrate the collision and then eventually verify
the collision problem has been fixed.

Signed-off-by: J Robert Ray <jrray@jrray.org>
Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>
Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>