Add storage generation concept
This offers a way to delay the use of v2 objects until time when spfs
can be updated everywhere that would access a repo with v2 objects in
it. Once the new spfs is seeded, the config can be updated to start using v2
objects. It is assumed that the config would be updated everywhere at
once.

Signed-off-by: J Robert Ray <jrray@jrray.org>