Write the platform first
If the blob is written first and an object with that digest didn't
actually exist yet, then the blob "wins" control of that digest, and
then creating the platform doesn't replace the digest file on disk,
keeping it a blob. In that case, the test passes even though a collision
problem still exists.

Signed-off-by: J Robert Ray <jrray@jrray.org>
Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>
Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>