Add a salt to digest calculation
With the previous changes, it is still relatively trivial to construct a
blob that matches empty platform, and it conceivable that it could
happen accidentally.

Adding the salt means the blob would have to also contain these bytes,
which would not happen accidentally. The uuid in the value is not
considered a secret.

Blobs themselves would benefit from also using a salt so that it would
be much more difficult to tailor make a blob to produce a specific
digest.

Signed-off-by: J Robert Ray <jrray@jrray.org>