Implement tag namespaces for rpc
This requires adding "*_in_namespace" flavors of the tag methods,
because the rpc server needed a way to process requests for tags in a
namespace without introducing race conditions by setting the tag
namespace on its repo handle. The server is "stateless" in the sense
that setting the tag namespace on the client side only changes state on
the client side, and then any requests sent to the server include the
name of the prevailing tag namespace as part of the request.

Signed-off-by: J Robert Ray <jrray@jrray.org>