Update graph objects to use flatbuffers
All top-level objects (that are constructed and stored on their own) are
now a specific variantion on an underlying generic type FlatObject.
These objects all contain the same AnyObject flatbuffer, with the more
specific types having validated the contained data as a specific
variant.

The main side effects of this are that:
 - a new Variant enum is created to enable the existing match-based
   disambiguation pattern
 - the spfs Repository trait can now have a generic write_object
   function that no longer requires wrapping the data in an emum before
   calling
 - objects must be constructed using the flatbuffer api, and types that
   are not top-level (tree, entry) cannot be created individually - only
   seen as references to data within the Manifest object

The removal of Tree and Entry as owned types also allows for removing
their references in other areas. They can no longer be encoded and saved
on their own, which was technically possible before but never done.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>
Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>