Fix test selectors interaction with components
Fixes #947.

This branch started off as a follow up to #816 because with that new
feature it became necessary for test selectors to be able to use the
same syntax. But in attempting to implement supporting the new syntax a
host of problems became apparent.

The first perhaps biggest problem was that although you could put some
expression like `"dep:comp1": "1.0.0"` in a variant, this wasn't being
seralized properly, so the version spec would end up with just
`"dep": "1.0.0"` and lose the component information.

Then, there are places where these "additional_requirements" in a
variant were simply being ignored.

Build digest calculation has been changed. The ability for OptionMap to
produce a digest has been removed. An `OptionMap` alone is not able to
calculate a correct digest, because `OptionMap`s can't describe package
requests with specific components enabled. The role of calculating
digests has been moved to `BuildSpec`.

The code for calculating digests the "proper" way (unlike how it was
being calculated during `BuildSpec` parsing) includes resolving options,
which requires knowing the package's name. Because of this, since deserializing
a `BuildSpec` means being able to validate that there are no duplicate variants,
and checking there are no duplicate variants requires calculating a digest, and
calculating a digest requires knowledge of the package name, a `BuildSpec`
cannot be deserialized straight from the recipe yaml anymore. There is now a
`TryFrom` implementation that can convert a tuple of
`(PkgName, UncheckedBuildSpec)` into a `BuildSpec`.

The behavior of test selectors has been changed to make them more
usable. Instead of expecting the defined selectors to produce a digest
that matches a variant in the recipe, they now act like a filter and
match on the variants. For example, specifying "python: 3" would match
any variants with "python: 3", rather than variants that only have
"python: 3" and nothing else.

Some random refactoring happened in the process of implementing these
changes, like `Digest` has been turned into a newtype struct, and the
`ComponentSet` code moved to its own module.

Signed-off-by: J Robert Ray <jrray@jrray.org>