Add test for expected behavior of host option vars with choices
The changes in the previous commit allow this new set of tests to pass.
The "choices" were being validated at BuildSpec deserialize time without
all the proper context available to validate them correctly.

The goal is to be able to read in a recipe that defines a host var with
choices but no default value.

Signed-off-by: J Robert Ray <jrray@jrray.org>