Make HOST_OPTIONS possible to be changed
In order to make tests that simulate HOST_OPTIONS with different values,
there needs to be a way to inject different values that will be returned
when accessing it.

`ArcSwap` was selected as an inexpensive way to access HOST_OPTIONS for
normal use and to have a way to make the value mutable for tests.

This implementation has been put behind a new type called `HostOptions`
in order to have the freedom to change the implementation later.

Signed-off-by: J Robert Ray <jrray@jrray.org>