Change spk Run trait to have generic return type
Rather than use i32 directly, allow this type be generic and only require
that it can be converted into an `ExitStatus`. The benefit may not be
clear yet, but this will allow tests to access more detailed information
about what happened after invoking `run()` on a cli tool implementation.

`ExitStatus` was used here instead of the more obvious choice of `ExitCode`
because `ExitCode` does not expose any method to inspect its value.

Signed-off-by: J Robert Ray <jrray@jrray.org>