Create types to hold build result information
This extends the idea of the build summary to collect the information in
a way that it can be utilized by test code, waiting to render it into a
string until the information is to be displayed.

The input "filename" is collected too, in case that is useful, since it
is possible to provide multiple files to build on the same command line.

Signed-off-by: J Robert Ray <jrray@jrray.org>