Repurpose the --variant flag
Refactor the variant selection logic so it can be reused, e.g., in the
cmd_test code.

Expand on `--variant` so it can be used to specify one or more variants
to build, either by index, or by passing in a json value.

The expected json value matches what could be placed inside a recipe as
an element in the variants list.

This logic requires access to the "raw" build.options value in the
recipe, which did not appear to already be accessible, so a new method
was added for that.

Signed-off-by: J Robert Ray <jrray@jrray.org>