Test behavior of --variant and --opt
Using --variant "overrules" (bypasses) the filtering mechanism, creating
a single build (per --variant arg) instead.

Whatever is specified via --opt is still respected. This leads to a
weird behavior of where a bespoke variant can be changed by --opt:

    spk build --variant '{ "color": "green" }' --opt 'color=blue'

This will build a package with color=blue. This behavior is what spills
out of the current code, and isn't necessarily what is desired. But the
alternative would be to ignore `--opt` in this case (or error?) which
are IMO worse outcomes, because they are not the least surprising
outcome.

If `--opt` is known to override values from variants already, then the
overrule behavior is predictable in this situation, even if it doesn't
make sense to use these flags in combination.

Signed-off-by: J Robert Ray <jrray@jrray.org>